fixed bug where error message displays on first render of loved products
